pyproject.toml
[project]
name = "carmannualqa"
version = "0.1.0"
description = "RAG-based chatbot for Mahindra Thar Car Manual"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "chromadb>=0.4.0",
    "python-dotenv>=1.0.0",
    "nltk>=3.8.0",
    "sentence-transformers>=2.2.0"
]
